+++ src/utils/api/fetch-ingredients.tsx
@@ -1,5 +1,5 @@
-import checkResponse from "./check-response";
-import { baseUrl, ingredientsUrl } from "./constants";
+import checkResponse from "../check-response";
+import { baseUrl, ingredientsUrl } from "../constants";
 
 export default function fetchIngredients (): Promise<any> {
   return fetch(`${baseUrl}${ingredientsUrl}`)
