+++ src/utils/api/register.js
@@ -1,5 +1,5 @@
-import checkResponse from "./check-response";
-import { baseUrl, registerUrl } from "./constants";
+import checkResponse from "../check-response";
+import { baseUrl, registerUrl } from "../constants";
 
 export default function register ({ username, email, password }) {
   return fetch(`${baseUrl}${registerUrl}`, {
